Add settings for CDS Hooks and use latest hapi-fhir-cr configs
