+++ src/main/java/ca/uhn/fhir/jpa/starter/cr/BaseCrConfig.java
@@ -1,17 +1,11 @@
 package ca.uhn.fhir.jpa.starter.cr;
 
-import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.cr.common.CodeCacheResourceChangeListener;
 import ca.uhn.fhir.cr.common.ElmCacheResourceChangeListener;
 import ca.uhn.fhir.jpa.api.dao.DaoRegistry;
-import ca.uhn.fhir.jpa.cache.IResourceChangeListenerCacheRefresher;
 import ca.uhn.fhir.jpa.cache.IResourceChangeListenerRegistry;
-import ca.uhn.fhir.jpa.cache.ResourceChangeListenerCacheFactory;
-import ca.uhn.fhir.jpa.cache.ResourceChangeListenerCacheRefresherImpl;
-import ca.uhn.fhir.jpa.cache.ResourceChangeListenerRegistryImpl;
 import ca.uhn.fhir.jpa.cache.ResourceChangeListenerRegistryInterceptor;
 import ca.uhn.fhir.jpa.searchparam.SearchParameterMap;
-import ca.uhn.fhir.jpa.searchparam.matcher.InMemoryResourceMatcher;
 import org.cqframework.cql.cql2elm.model.CompiledLibrary;
 import org.cqframework.cql.cql2elm.model.Model;
 import org.hl7.cql.model.ModelIdentifier;
@@ -28,6 +22,11 @@ import java.util.concurrent.ConcurrentHashMap;
 @Configuration
 public class BaseCrConfig {
 	@Bean
+	public CrProperties crProperties() {
+		return new CrProperties();
+	}
+
+	@Bean
 	public Map<VersionedIdentifier, CompiledLibrary> globalLibraryCache() {
 		return new ConcurrentHashMap<>();
 	}
@@ -41,7 +40,6 @@ public class BaseCrConfig {
 	public Map<String, List<Code>> globalValueSetCache() {
 		return new ConcurrentHashMap<>();
 	}
-
 
 	@Bean
 	public ElmCacheResourceChangeListener elmCacheResourceChangeListener(
@@ -69,15 +67,17 @@ public class BaseCrConfig {
 		return listener;
 	}
 
-	@Bean
-	public IResourceChangeListenerRegistry resourceChangeListenerRegistry(InMemoryResourceMatcher theInMemoryResourceMatcher, FhirContext theFhirContext, ResourceChangeListenerCacheFactory theResourceChangeListenerCacheFactory) {
-		return new ResourceChangeListenerRegistryImpl(theFhirContext, theResourceChangeListenerCacheFactory, theInMemoryResourceMatcher);
-	}
+	// These beans were being duplicated
+	// @Bean
+	// public IResourceChangeListenerRegistry resourceChangeListenerRegistry(InMemoryResourceMatcher theInMemoryResourceMatcher, FhirContext theFhirContext, ResourceChangeListenerCacheFactory theResourceChangeListenerCacheFactory) {
+	// 	return new ResourceChangeListenerRegistryImpl(theFhirContext, theResourceChangeListenerCacheFactory, theInMemoryResourceMatcher);
+	// }
 
-	@Bean
-	IResourceChangeListenerCacheRefresher resourceChangeListenerCacheRefresher() {
-		return new ResourceChangeListenerCacheRefresherImpl();
-	}
+	// @Bean
+	// IResourceChangeListenerCacheRefresher resourceChangeListenerCacheRefresher() {
+	// 	return new ResourceChangeListenerCacheRefresherImpl();
+	// }
+
 	@Bean
 	public ResourceChangeListenerRegistryInterceptor resourceChangeListenerRegistryInterceptor() {
 		return new ResourceChangeListenerRegistryInterceptor();
