+++ src/main/java/ca/uhn/fhir/jpa/starter/cr/StarterCrR4Config.java
@@ -1,10 +1,15 @@
 package ca.uhn.fhir.jpa.starter.cr;
 
+import ca.uhn.fhir.cr.config.r4.ApplyOperationConfig;
+import ca.uhn.fhir.cr.config.r4.CrR4Config;
+import ca.uhn.fhir.cr.config.r4.ExtractOperationConfig;
+import ca.uhn.fhir.cr.config.r4.PackageOperationConfig;
+import ca.uhn.fhir.cr.config.r4.PopulateOperationConfig;
 import ca.uhn.fhir.jpa.starter.annotations.OnR4Condition;
 import ca.uhn.fhir.rest.server.RestfulServer;
 import ca.uhn.fhir.rest.server.provider.ResourceProviderFactory;
 import ca.uhn.fhir.cr.common.CqlThreadFactory;
-import ca.uhn.fhir.jpa.starter.AppProperties;
+
 import org.cqframework.cql.cql2elm.CqlCompilerOptions;
 import org.cqframework.cql.cql2elm.model.CompiledLibrary;
 import org.cqframework.cql.cql2elm.model.Model;
@@ -18,7 +23,11 @@ import org.opencds.cqf.fhir.cr.measure.MeasureEvaluationOptions;
 import org.opencds.cqf.fhir.utility.ValidationProfile;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.context.annotation.*;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Conditional;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+import org.springframework.context.annotation.Primary;
 import org.springframework.security.concurrent.DelegatingSecurityContextExecutorService;
 
 import java.util.EnumSet;
@@ -30,7 +39,14 @@ import java.util.concurrent.Executors;
 
 @Configuration
 @Conditional({ OnR4Condition.class, CrConfigCondition.class })
-@Import({BaseCrConfig.class, CrR4Config.class})
+@Import({
+	BaseCrConfig.class,
+	CrR4Config.class,
+	ApplyOperationConfig.class,
+	ExtractOperationConfig.class,
+	PackageOperationConfig.class,
+	PopulateOperationConfig.class
+})
 public class StarterCrR4Config {
 	private static final Logger ourLogger = LoggerFactory.getLogger(StarterCrR4Config.class);
 
@@ -47,11 +63,11 @@ public class StarterCrR4Config {
 	}
 
 	@Bean
-	CareGapsProperties careGapsProperties(AppProperties theAppProperties)  {
+	CareGapsProperties careGapsProperties(CrProperties theCrProperties)  {
 		var careGapsProperties = new CareGapsProperties();
 		careGapsProperties.setThreadedCareGapsEnabled(false);
-		careGapsProperties.setCareGapsReporter(theAppProperties.getCareGapsReporter());
-		careGapsProperties.setCareGapsCompositionSectionAuthor(theAppProperties.getCareGapsSectionAuthor());
+		careGapsProperties.setCareGapsReporter(theCrProperties.getCareGapsReporter());
+		careGapsProperties.setCareGapsCompositionSectionAuthor(theCrProperties.getCareGapsSectionAuthor());
 		return careGapsProperties;
 	}
 
@@ -67,7 +83,7 @@ public class StarterCrR4Config {
 
 	@Bean
 	public EvaluationSettings evaluationSettings(
-		AppProperties theAppProperties,
+		CrProperties theCrProperties,
 		Map<VersionedIdentifier, CompiledLibrary> theGlobalLibraryCache,
 		Map<ModelIdentifier, Model> theGlobalModelCache,
 		Map<String, List<Code>> theGlobalValueSetCache) {
@@ -76,10 +92,10 @@ public class StarterCrR4Config {
 
 		var cqlEngineOptions = cqlOptions.getCqlEngineOptions();
 		Set<CqlEngine.Options> options = EnumSet.noneOf(CqlEngine.Options.class);
-		if (theAppProperties.isCqlRuntimeEnableExpressionCaching()) {
+		if (theCrProperties.isCqlRuntimeEnableExpressionCaching()) {
 			options.add(CqlEngine.Options.EnableExpressionCaching);
 		}
-		if (theAppProperties.isCqlRuntimeEnableValidation()) {
+		if (theCrProperties.isCqlRuntimeEnableValidation()) {
 			options.add(CqlEngine.Options.EnableValidation);
 		}
 		cqlEngineOptions.setOptions(options);
@@ -87,53 +103,52 @@ public class StarterCrR4Config {
 
 		var cqlCompilerOptions = new CqlCompilerOptions();
 
-		if (theAppProperties.isEnableDateRangeOptimization()
-		) {
+		if (theCrProperties.isEnableDateRangeOptimization()) {
 			cqlCompilerOptions.setOptions(CqlCompilerOptions.Options.EnableDateRangeOptimization);
 		}
-		if (theAppProperties.isEnableAnnotations()) {
+		if (theCrProperties.isEnableAnnotations()) {
 			cqlCompilerOptions.setOptions(CqlCompilerOptions.Options.EnableAnnotations);
 		}
-		if (theAppProperties.isEnableLocators()) {
+		if (theCrProperties.isEnableLocators()) {
 			cqlCompilerOptions.setOptions(CqlCompilerOptions.Options.EnableLocators);
 		}
-		if (theAppProperties.isEnableResultsType()) {
+		if (theCrProperties.isEnableResultsType()) {
 			cqlCompilerOptions.setOptions(CqlCompilerOptions.Options.EnableResultTypes);
 		}
-		cqlCompilerOptions.setVerifyOnly(theAppProperties.isCqlCompilerVerifyOnly());
-		if (theAppProperties.isEnableDetailedErrors()) {
+		cqlCompilerOptions.setVerifyOnly(theCrProperties.isCqlCompilerVerifyOnly());
+		if (theCrProperties.isEnableDetailedErrors()) {
 			cqlCompilerOptions.setOptions(CqlCompilerOptions.Options.EnableDetailedErrors);
 		}
-		cqlCompilerOptions.setErrorLevel(theAppProperties.getCqlCompilerErrorSeverityLevel());
-		if (theAppProperties.isDisableListTraversal()) {
+		cqlCompilerOptions.setErrorLevel(theCrProperties.getCqlCompilerErrorSeverityLevel());
+		if (theCrProperties.isDisableListTraversal()) {
 			cqlCompilerOptions.setOptions(CqlCompilerOptions.Options.DisableListTraversal);
 		}
-		if (theAppProperties.isDisableListDemotion()) {
+		if (theCrProperties.isDisableListDemotion()) {
 			cqlCompilerOptions.setOptions(CqlCompilerOptions.Options.DisableListDemotion);
 		}
-		if (theAppProperties.isDisableListPromotion()) {
+		if (theCrProperties.isDisableListPromotion()) {
 			cqlCompilerOptions.setOptions(CqlCompilerOptions.Options.DisableListPromotion);
 		}
-		if (theAppProperties.isEnableIntervalDemotion()) {
+		if (theCrProperties.isEnableIntervalDemotion()) {
 			cqlCompilerOptions.setOptions(CqlCompilerOptions.Options.EnableIntervalDemotion);
 		}
-		if (theAppProperties.isEnableIntervalPromotion()) {
+		if (theCrProperties.isEnableIntervalPromotion()) {
 			cqlCompilerOptions.setOptions(CqlCompilerOptions.Options.EnableIntervalPromotion);
 		}
-		if (theAppProperties.isDisableMethodInvocation()) {
+		if (theCrProperties.isDisableMethodInvocation()) {
 			cqlCompilerOptions.setOptions(CqlCompilerOptions.Options.DisableMethodInvocation);
 		}
-		if (theAppProperties.isRequireFromKeyword()) {
+		if (theCrProperties.isRequireFromKeyword()) {
 			cqlCompilerOptions.setOptions(CqlCompilerOptions.Options.RequireFromKeyword);
 		}
-		cqlCompilerOptions.setValidateUnits(theAppProperties.isCqlCompilerValidateUnits());
-		if (theAppProperties.isDisableDefaultModelInfoLoad()) {
+		cqlCompilerOptions.setValidateUnits(theCrProperties.isCqlCompilerValidateUnits());
+		if (theCrProperties.isDisableDefaultModelInfoLoad()) {
 			cqlCompilerOptions.setOptions(CqlCompilerOptions.Options.DisableDefaultModelInfoLoad);
 		}
-		cqlCompilerOptions.setSignatureLevel(theAppProperties.getCqlCompilerSignatureLevel());
-		cqlCompilerOptions.setCompatibilityLevel(theAppProperties.getCqlCompilerCompatibilityLevel());
-		cqlCompilerOptions.setAnalyzeDataRequirements(theAppProperties.isCqlCompilerAnalyzeDataRequirements());
-		cqlCompilerOptions.setCollapseDataRequirements(theAppProperties.isCqlCompilerCollapseDataRequirements());
+		cqlCompilerOptions.setSignatureLevel(theCrProperties.getCqlCompilerSignatureLevel());
+		cqlCompilerOptions.setCompatibilityLevel(theCrProperties.getCqlCompilerCompatibilityLevel());
+		cqlCompilerOptions.setAnalyzeDataRequirements(theCrProperties.isCqlCompilerAnalyzeDataRequirements());
+		cqlCompilerOptions.setCollapseDataRequirements(theCrProperties.isCqlCompilerCollapseDataRequirements());
 
 		cqlOptions.setCqlCompilerOptions(cqlCompilerOptions);
 		evaluationSettings.setLibraryCache(theGlobalLibraryCache);
