+++ src/main/resources/application.yaml
@@ -53,14 +53,19 @@ spring:
 #      hibernate.search.backend.analysis.configurer: ca.uhn.fhir.jpa.search.HapiHSearchAnalysisConfigurers$HapiElasticAnalysisConfigurer
 hapi:
   fhir:
+    ### This flag when enabled to true, will avail evaluate measure operations from CR Module.
+    ### Flag is false by default, can be passed as command line argument to override.
+    cr:
+      enabled: true
+
+    cdshooks:
+      enabled: true
+      clientIdHeaderName: client_id
 
     ### This enables the swagger-ui at /fhir/swagger-ui/index.html as well as the /fhir/api-docs (see https://hapifhir.io/hapi-fhir/docs/server_plain/openapi.html)
     openapi_enabled: true
     ### This is the FHIR version. Choose between, DSTU2, DSTU3, R4 or R5
     fhir_version: R4
-    ### This flag when enabled to true, will avail evaluate measure operations from CR Module.
-    ### Flag is false by default, can be passed as command line argument to override.
-    cr_enabled: true
 
     ### enable to use the ApacheProxyAddressStrategy which uses X-Forwarded-* headers
     ### to determine the FHIR server address
@@ -104,7 +109,6 @@ hapi:
     #    auto_create_placeholder_reference_targets: false
     ### tells the server to automatically append the current version of the target resource to references at these paths
     #    auto_version_reference_at_paths: Device.patient, Device.location, Device.parent, DeviceMetric.parent, DeviceMetric.source, Observation.device, Observation.subject
-    #    cr_enabled: true
     #    ips_enabled: false
     #    default_encoding: JSON
     #    default_pretty_print: true
